% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{ceil_digit}
\alias{ceil_digit}
\title{ceil_digit}
\usage{
ceil_digit(x, n)
}
\arguments{
\item{x}{A numeric vector to round up.}

\item{n}{Integer. Number of decimal places to retain.}
}
\value{
A numeric vector rounded up to `n` digits.
}
\description{
Round Up a Number to a Specified Number of Decimal Places
}
\details{
Rounds numeric values upward to the nearest multiple at `n` decimal places.
}
\examples{
ceil_digit(3.1415, 2)  # 3.15
ceil_digit(c(2.001, 2.009), 2)  # c(2.01, 2.01)

}
