% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coembed.R
\name{cell_knn_matrix_mutiplication}
\alias{cell_knn_matrix_mutiplication}
\title{cell_knn_matrix_mutiplication}
\usage{
cell_knn_matrix_mutiplication(knn, cell_feature_mat, feature_feature_mat)
}
\arguments{
\item{knn}{A matrix or data frame with two columns: `i` (source cell) and `j` (neighbor cell).}

\item{cell_feature_mat}{A numeric matrix of cell-by-feature values.}

\item{feature_feature_mat}{A matrix representing relationships between features (e.g., clone similarity).}
}
\value{
A numeric vector of scores, one per kNN edge.
}
\description{
Matrix Multiplication for Cell-Feature Weighted Edges
}
\details{
Computes a score for each kNN edge using the dot product between a cell’s features and a transformed clone matrix.
Used to integrate clone-to-clone and cell-to-clone relationships into edge-level scores.
}
