% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{cluster_merge}
\alias{cluster_merge}
\title{cluster_merge}
\usage{
cluster_merge(input_list, cluster)
}
\arguments{
\item{input_list}{A list of vectors or objects to be grouped.}

\item{cluster}{A vector of cluster labels of the same length as `input_list`.}
}
\value{
A list where each element contains the merged values from the same cluster.
}
\description{
Merge Elements from a List According to Cluster Assignments
}
\details{
Given a list and a vector of cluster assignments, merges list elements that belong to the same cluster into grouped lists.
}
\examples{
lst <- list(1:2, 3:4, 5, 6)
clust <- c(1, 1, 2, 2)
merged <- cluster_merge(lst, clust)
print(merged)

}
