% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{mat_split}
\alias{mat_split}
\title{mat_split}
\usage{
mat_split(mat, size, index = 0)
}
\arguments{
\item{mat}{A matrix to be split.}

\item{size}{An integer vector specifying the size of each chunk (rows or columns).}

\item{index}{Integer. If `0` (default), splits by rows; if `1`, splits by columns.}
}
\value{
A list of matrices representing the chunks.
}
\description{
Split a Matrix into Row or Column Chunks
}
\details{
Splits a matrix into smaller matrices based on user-defined row or column sizes.
}
\examples{
mat <- matrix(1:20, nrow = 5)
chunks <- mat_split(mat, size = c(2, 3), index = 0)
str(chunks)

}
