% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{clone_root}
\alias{clone_root}
\title{clone_root}
\usage{
clone_root(clones, cell_meta, clone_col, cluster_col, start_cluster)
}
\arguments{
\item{clones}{A character vector of clone identifiers.}

\item{cell_meta}{A data frame with clone and cluster information per cell.}

\item{clone_col}{Name of the column containing clone labels.}

\item{cluster_col}{Name of the column containing cluster labels.}

\item{start_cluster}{Name or value of the cluster to define the root state.}
}
\value{
The name of the most enriched clone in the start cluster.
}
\description{
Identify Root Clone from Cluster Enrichment
}
\details{
Finds the most enriched clone in a given starting cluster, to use as a pseudotime root.
}
\examples{
root_clone <- clone_root(clones = c("A", "B", "C"),
                         cell_meta = cell_metadata,
                         clone_col = "clone_id",
                         cluster_col = "cluster",
                         start_cluster = "Naive")

}
