% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{DPT_T}
\alias{DPT_T}
\title{DPT_T}
\usage{
DPT_T(T_mat, start)
}
\arguments{
\item{T_mat}{A row-normalized transition matrix (symmetric or stochastic).}

\item{start}{Integer. Index of the root cell for pseudotime computation.}
}
\value{
A numeric vector of diffusion pseudotime values for all nodes.
}
\description{
Compute Diffusion Pseudotime from Accumulated Commute Time Matrix
}
\details{
Computes diffusion pseudotime (DPT) distances from a specified root node using the accumulated commute time matrix.
}
\examples{
T <- matrix(runif(100), 10, 10)
T <- T / rowSums(T)
pt <- DPT_T(T, start = 1)

}
