% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{link2cluster}
\alias{link2cluster}
\title{link2cluster}
\usage{
link2cluster(link, nodes)
}
\arguments{
\item{link}{A data frame with two columns (i, j) indicating node pairs with links.}

\item{nodes}{A character or factor vector of all node names to be clustered.}
}
\value{
A numeric vector of cluster labels for each node, as assigned by DBSCAN.
}
\description{
Cluster Nodes Based on Link Connectivity
}
\details{
Given a set of links between node pairs and a list of all nodes, this function constructs a symmetric adjacency matrix,
applies diffusion to expand connectivity, and clusters the result using DBSCAN.
}
\examples{
link_df <- data.frame(i = c("A", "B", "C"), j = c("B", "C", "A"))
clusters <- link2cluster(link_df, nodes = c("A", "B", "C"))
print(clusters)

}
